Reorg migration and pallet and pallet impls
